# fix: lower string literals with actual value instead of default