Fixes #11582. Some questions below in the code review section.