@tysg 2022-03-04
Did some memory profiling using `cargo run --bin rust-analyzer --release -- -q analysis-stats --memory-usage .`

Before (9d473a0b9f92c0739e4635fa8ef95a91ac2a481d):
```
Running `target/release/rust-analyzer -q analysis-stats --memory-usage .`
Failed to create perf counter: Operation not permitted (os error 1)
Failed to create perf counter: Operation not permitted (os error 1)
Database loaded:     392.17ms, 98mb (metadata 222.96ms, 489kb; build 68.66ms, 43kb)
Failed to create perf counter: Operation not permitted (os error 1)
  crates: 39, mods: 854, decls: 19304, fns: 14437
Item Collection:     8.08s, 352mb
Failed to create perf counter: Operation not permitted (os error 1)
  exprs: 397281, ??ty: 336 (0%), ?ty: 158 (0%), !ty: 144
Inference:           41.06s, 683mb
Total:               49.14s, 1035mb
```

After:
```
Failed to create perf counter: Operation not permitted (os error 1)
Failed to create perf counter: Operation not permitted (os error 1)
Database loaded:     9.38s, 98mb (metadata 224.45ms, 489kb; build 9.05s, 46kb)
Failed to create perf counter: Operation not permitted (os error 1)
  crates: 39, mods: 854, decls: 19304, fns: 14437
Item Collection:     8.01s, 352mb
Failed to create perf counter: Operation not permitted (os error 1)
  exprs: 397297, ??ty: 336 (0%), ?ty: 163 (0%), !ty: 144
Inference:           42.21s, 684mb
Total:               50.22s, 1037mb
```
Looks like there's no significant change in memory usage as I understand. 

Regarding `Byte` and `Char`, since the `LiteralKind` intermediate doesn't record the value of the Byte and Char literals, we can't fix it like this. I don't think using the `LiteralKind` intermediate is the best way to lower an AST to HIR (I asked [a question about this on Zulip](https://rust-lang.zulipchat.com/#narrow/stream/185405-t-compiler.2Frust-analyzer/topic/lowering.3A.20ast.3A.3ALiteral.20-.3E.20LiteralKind.20-.3E.20hir.3A.3ALiteral)). Should the team think so too, I'll send in the changes in a separate PR.

So this PR is ready for merge/review now @lnicola, thanks!